+++ kafka-common/src/main/java/com/networknt/kafka/common/AbstractAvroSerDe.java
@@ -1,9 +1,6 @@
 package com.networknt.kafka.common;
 
 import com.networknt.config.Config;
-import com.networknt.kafka.common.config.KafkaConsumerConfig;
-import com.networknt.kafka.common.config.KafkaProducerConfig;
-import com.networknt.kafka.common.config.KafkaStreamsConfig;
 import com.networknt.service.SingletonServiceFactory;
 import io.confluent.kafka.schemaregistry.client.SchemaRegistryClient;
