+++ kafka-common/src/test/java/com/networknt/kafka/common/KafkaConsumerConfigTest.java
@@ -1,6 +1,5 @@
 package com.networknt.kafka.common;
 
-import com.networknt.kafka.common.config.KafkaConsumerConfig;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
