+++ kafka-common/src/test/java/com/networknt/kafka/common/KafkaProducerConfigTest.java
@@ -1,6 +1,5 @@
 package com.networknt.kafka.common;
 
-import com.networknt.kafka.common.config.KafkaProducerConfig;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
