+++ kafka-common/src/test/java/com/networknt/kafka/common/KafkaStreamsConfigTest.java
@@ -1,6 +1,5 @@
 package com.networknt.kafka.common;
 
-import com.networknt.kafka.common.config.KafkaStreamsConfig;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
