+++ kafka-consumer/src/main/java/com/networknt/kafka/consumer/AbstractConsumer.java
@@ -1,7 +1,7 @@
 package com.networknt.kafka.consumer;
 
 import com.networknt.config.Config;
-import com.networknt.kafka.common.config.KafkaConsumerConfig;
+import com.networknt.kafka.common.KafkaConsumerConfig;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
