+++ kafka-consumer/src/main/java/com/networknt/kafka/consumer/KafkaConsumerManager.java
@@ -18,10 +18,9 @@ package com.networknt.kafka.consumer;
 import com.networknt.client.Http2Client;
 import com.networknt.config.JsonMapper;
 import com.networknt.exception.FrameworkException;
-import com.networknt.kafka.common.config.KafkaConsumerConfig;
+import com.networknt.kafka.common.KafkaConsumerConfig;
 import com.networknt.kafka.entity.*;
 import com.networknt.status.Status;
-import com.networknt.utility.Constants;
 import com.networknt.utility.Constants;
 import com.networknt.utility.StringUtils;
 import io.undertow.client.ClientConnection;
