+++ kafka-consumer/src/main/java/com/networknt/kafka/consumer/LightConsumer.java
@@ -1,10 +1,5 @@
 package com.networknt.kafka.consumer;
 
-import com.networknt.kafka.common.config.KafkaConsumerConfig;
-
-import java.util.ArrayList;
-import java.util.List;
-
 public interface LightConsumer {
     void open();
     void close();
