+++ kafka-producer/src/main/java/com/networknt/kafka/producer/TransactionalProducer.java
@@ -2,7 +2,7 @@ package com.networknt.kafka.producer;
 
 import com.networknt.config.Config;
 import com.networknt.httpstring.HttpStringConstants;
-import com.networknt.kafka.common.config.KafkaProducerConfig;
+import com.networknt.kafka.common.KafkaProducerConfig;
 import com.networknt.kafka.common.TransactionalKafkaException;
 import com.networknt.kafka.common.FlinkKafkaProducer;
 import com.networknt.server.ServerConfig;
